+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,133 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.85.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.85.2...@coveo/quantic@0.85.3) (2021-11-14)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.85.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.85.1...@coveo/quantic@0.85.2) (2021-11-14)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.85.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.85.0...@coveo/quantic@0.85.1) (2021-11-12)
+
+
+### Bug Fixes
+
+* **quantic:** copy only main headless browser file ([#1480](https://github.com/coveo/ui-kit/issues/1480)) ([11a1549](https://github.com/coveo/ui-kit/commit/11a15498f337f99fd6f2191608312a2387aeaa6c))
+
+
+
+
+
+# [0.85.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.84.2...@coveo/quantic@0.85.0) (2021-11-10)
+
+
+### Features
+
+* **quantic:** fixed facet count alignment ([#1467](https://github.com/coveo/ui-kit/issues/1467)) ([ae83e54](https://github.com/coveo/ui-kit/commit/ae83e54386c0c04940323dbb9ad2d8a89cf9d2c8))
+
+
+
+
+
+## [0.84.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.84.1...@coveo/quantic@0.84.2) (2021-11-10)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.84.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.84.0...@coveo/quantic@0.84.1) (2021-11-10)
+
+
+### Bug Fixes
+
+* **quantic:** text truncating class returned to container in template ([#1464](https://github.com/coveo/ui-kit/issues/1464)) ([5c4a0af](https://github.com/coveo/ui-kit/commit/5c4a0af951effadbab934a7c1e480abfc15db2d7))
+
+
+
+
+
+# [0.84.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.83.2...@coveo/quantic@0.84.0) (2021-11-09)
+
+
+### Bug Fixes
+
+* **quantic:** hide timeframe facet when results do not match the specified timeframes ([#1459](https://github.com/coveo/ui-kit/issues/1459)) ([eebf443](https://github.com/coveo/ui-kit/commit/eebf44305d928eaca82928c146d98084668e1988))
+
+
+### Features
+
+* **quantic:** removed delimitingCharacter ([#1458](https://github.com/coveo/ui-kit/issues/1458)) ([67c5f91](https://github.com/coveo/ui-kit/commit/67c5f91170e3713193c3c8bcae3ba99f47dde80f))
+
+
+
+
+
+## [0.83.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.83.1...@coveo/quantic@0.83.2) (2021-11-09)
+
+
+### Bug Fixes
+
+* **quantic:** added result-template sections and icon maps ([#1442](https://github.com/coveo/ui-kit/issues/1442)) ([40ffb83](https://github.com/coveo/ui-kit/commit/40ffb838f76808ea1e12c6699fbeaa02d72916bd))
+* **quantic:** added truncate and removed placeholder min-width ([#1452](https://github.com/coveo/ui-kit/issues/1452)) ([4beeec4](https://github.com/coveo/ui-kit/commit/4beeec487eb9f2bce6aec157606afd5240b00da6))
+* **quantic:** fixed validation bug with numeric facet ([#1457](https://github.com/coveo/ui-kit/issues/1457)) ([d1bca6e](https://github.com/coveo/ui-kit/commit/d1bca6e5b36374ae425e4c850b45967d79536dab))
+* **quantic:** remove timeframe facet from community builder ([#1451](https://github.com/coveo/ui-kit/issues/1451)) ([ddc9c6c](https://github.com/coveo/ui-kit/commit/ddc9c6c79aebd9b2f204712d4ce220732b9c5314))
+* **quantic:** removed onchange handlers causing error ([#1454](https://github.com/coveo/ui-kit/issues/1454)) ([aebb895](https://github.com/coveo/ui-kit/commit/aebb895b2dfdc90ddb11dc379bede7399033e776))
+* **quantic:** various timeframe facet issues ([#1456](https://github.com/coveo/ui-kit/issues/1456)) ([9df126c](https://github.com/coveo/ui-kit/commit/9df126c167f3b8b54f621dff5512bc8acfc4b85d))
+
+
+
+
+
+## [0.83.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.83.0...@coveo/quantic@0.83.1) (2021-11-08)
+
+
+### Bug Fixes
+
+* **quantic:** fix facet manager issue in local dev server ([#1450](https://github.com/coveo/ui-kit/issues/1450)) ([7f0349e](https://github.com/coveo/ui-kit/commit/7f0349e6323f5bd61f115f5093b566d4686bd2c2))
+
+
+
+
+
+# [0.83.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.82.0...@coveo/quantic@0.83.0) (2021-11-08)
+
+
+### Bug Fixes
+
+* **quantic:** support facet with child elements in facet manager ([#1443](https://github.com/coveo/ui-kit/issues/1443)) ([9842a2e](https://github.com/coveo/ui-kit/commit/9842a2ea3e61a0184db6622e008e1ce8f89c3964))
+
+
+### Features
+
+* **quantic:** add timeframe facet ([#1438](https://github.com/coveo/ui-kit/issues/1438)) ([d31aab6](https://github.com/coveo/ui-kit/commit/d31aab692230899bc1601ac16bc40ce4e5d6caba))
+
+
+
+
+
+# [0.82.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.81.0...@coveo/quantic@0.82.0) (2021-11-05)
+
+
+### Features
+
+* **quantic:** added ResultList E2E tests ([#1428](https://github.com/coveo/ui-kit/issues/1428)) ([211af9b](https://github.com/coveo/ui-kit/commit/211af9bb2d56444c8e013f456fe70eeebb6a1a21))
+
+
+
+
+
 # [0.81.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.80.7...@coveo/quantic@0.81.0) (2021-11-03)